requirements/base.txt
# Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.0

# Authentication & Security
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Video Processing & AI
moviepy==1.0.3
opencv-python==4.8.1.78
mediapipe==0.10.8
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
librosa==0.10.1
numpy==1.24.3
soundfile==0.12.1
Pillow>=10.3.0

# File Storage
boto3==1.34.0  # AWS S3

# Payment
stripe==7.8.0

# Email
sendgrid==6.11.0

# Monitoring & Logging
sentry-sdk[flask]==1.39.1
prometheus-client==0.19.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
marshmallow==3.20.1
validators==0.22.0

# Media Utils
imageio>=2.34.0
imageio-ffmpeg==0.4.9
ffmpeg-python==0.2.0
